Move WhatsNew to components/index
