+++ pages/index.vue
@@ -52,8 +52,8 @@ import Vue from 'vue'
 import { MetaInfo } from 'vue-meta'
 
 import PageHeader from '@/components/PageHeader.vue'
-import WhatsNew from '@/components/WhatsNew.vue'
 
+import WhatsNew from '@/components/index/WhatsNew.vue'
 import StaticInfo from '@/components/index/StaticInfo.vue'
 import CardRow from '@/components/index/CardRow.vue'
